Update Dockerfile
Now points to specific versions of biopython and numpy, both of which are soon to eschew python2, which is what the included version of the CARD/RGI tools is using.  Also moved Bio::SeqIO to the end of the list of perl modules installed by cpanm, which somehow resolved some errors I was getting while building the docker image.